News cache and move files for MVVM
What's new:
- News cache for faster loading and offline use.
- Move files to follow MVVM
