+++ app/src/main/java/io/zoemeow/dutapp/di/NewsCacheDBModule.kt
@@ -7,32 +7,14 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import io.zoemeow.dutapp.data.NewsCacheDatabase
-import io.zoemeow.dutapp.data.NewsGlobalCacheDatabaseDao
-import io.zoemeow.dutapp.data.NewsSubjectCacheDatabaseDao
-import io.zoemeow.dutapp.network.DutFuncApi
-import io.zoemeow.dutapp.repository.DutNewsRepository
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
+import io.zoemeow.dutapp.data.db.NewsCacheDatabase
+import io.zoemeow.dutapp.data.db.NewsGlobalCacheDatabaseDao
+import io.zoemeow.dutapp.data.db.NewsSubjectCacheDatabaseDao
 import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-object AppModule {
-    @Singleton
-    @Provides
-    fun provideNewsRepository(api: DutFuncApi) = DutNewsRepository(api)
-
-    @Singleton
-    @Provides
-    fun provideDutFuncApi(): DutFuncApi {
-        return Retrofit.Builder()
-            .baseUrl("https://dutapi.herokuapp.com")
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(DutFuncApi::class.java)
-    }
-
+object NewsCacheDBModule {
     // Database for News global cache and news subject cache
     @Provides
     @Singleton
