--- app/src/main/java/io/zoemeow/dutapp/navbar/NavLoginRoutes.kt
@@ -1,7 +1,0 @@
-package io.zoemeow.dutapp.navbar
-
-sealed class NavLoginRoutes(val route: String) {
-    object AccountPageNotLoggedIn : NavLoginRoutes("notloggedin")
-    object AccountPageLogin : NavLoginRoutes("login")
-    object AccountPageLoggedIn : NavLoginRoutes("loggedin")
-}
