+++ app/src/main/java/io/zoemeow/dutapp/utils/LinkItemConverter.kt
@@ -1,4 +1,4 @@
-package io.zoemeow.dutapp
+package io.zoemeow.dutapp.utils
 
 import androidx.room.TypeConverter
 import com.google.gson.Gson
